requirements.txt
asttokens==2.0.8
backcall==0.2.0
debugpy==1.6.3
decorator==5.1.1
entrypoints==0.4
executing==1.0.0
ipykernel==6.15.2
ipython==8.5.0
jedi==0.18.1
jupyter-core==4.11.1
jupyter_client==7.3.5
matplotlib-inline==0.1.6
nest-asyncio==1.5.5
packaging==21.3
parso==0.8.3
pexpect==4.8.0
pickleshare==0.7.5
prompt-toolkit==3.0.31
psutil==5.9.2
ptyprocess==0.7.0
pure-eval==0.2.2
py4j==0.10.9.5
Pygments==2.13.0
pyparsing==3.0.9
pyspark==3.3.0
python-dateutil==2.8.2
pyzmq==23.2.1
six==1.16.0
stack-data==0.5.0
tornado==6.2
traitlets==5.3.0
wcwidth==0.2.5
